requirements.txt
altair
pandas
streamlit
torch
transformers
Pillow
requests
pdf2image